Dockerfile and compose.yaml (and settings.DATABASES)
This sets up a consistent set of environment variables and
default values for local TOM development (i.e. runserver) with
a dockerized postgres database (see comments in settings.py).
Also, the environment variable in the compose.yaml are consistent
witht the os.getenv calls in the settings.DATABASES['default']
configuration. The main point is that the settings.py, Dockerfile,
and compose.yaml are interdependent and working together.